Package refactoring for UserController. Removed unused imports
