--- src/main/java/com/kodilla/ecommercee/controller/UserController.java
@@ -1,25 +1,0 @@
-package com.kodilla.ecommercee.controller;
-
-
-import com.kodilla.ecommercee.user.domain.UserDto;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("ecommercee/users")
-public class UserController {
-    @PostMapping
-    public Long createUser(@RequestBody String userName) {
-        UserDto userDto = new UserDto(1L, userName, "1", 12345L);
-        return userDto.getUserId();
-    }
-
-    @PutMapping(path = "/blockUser")
-    public Long blockUser(@RequestParam Long userId) {
-        return userId;
-    }
-
-    @GetMapping(path = "/generateKey")
-    public Long generateKey(@RequestParam Long userId) {
-        return 12345678L;
-    }
-}
