+++ src/main/java/com/kodilla/ecommercee/user/domain/UserDto.java
@@ -1,7 +1,6 @@
 package com.kodilla.ecommercee.user.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
